+++ Logic/ViewModel/CarViewModel.cs
@@ -18,16 +18,16 @@ namespace Logic.ViewModel
 
         [DataMember]
         [DisplayName("Цена автомобиля без комплектации")]
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
         [DataMember]
         [DisplayName("Цена автомобиля")]
-        public decimal FullPrice { get; set; }
+        public int FullPrice { get; set; }
 
         [DataMember]
         [DisplayName("Год выпуска")]
         public int Year { get; set; }
 
-        public Dictionary<int, (string, int, decimal)> CarDetails { get; set; }
+        public Dictionary<int, (string, int, int)> CarDetails { get; set; }
     }
 }
